GeneratePoints

Description

Documentation

generatePoints :: (Double, Double) -> Double -> Double -> String -> Double -> [(Double, Double)]

Function generatePoints generates list of points based on String argument where first argument is initial position (point), the second is startAngle in radians, the third is changeAngle in radians which refers to + and - characters in String argument, the fourth is String which contains only F, + and - characters, the fifth is n which is the length of the line (distance between two points in list)