GenerateString

Description

Documentation

generateSingle :: String -> String -> String

Function generateSingle generates String by replacing every F character in first argument firstString by secondString.

generateFull :: String -> Integer -> String -> String

Function generateFull generates String by replacing every F character in firstString by secondString n times (iterations), which means that in each iteration firstString is a String generated from previous iteration.